+++ app/api/brief/[cid]/route.ts
@@ -1,24 +1,8 @@
 import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
-import { NextRequest, NextResponse } from "next/server";
-import { getFileEntry } from "@/api/dbChartFile";
-import { ChartBrief } from "@/chartFormat/chart";
+import { NextRequest } from "next/server";
+import { getBrief } from "../brief";
 
 export async function GET(request: NextRequest, context: { params: Params }) {
   const cid: string = context.params.cid;
   return getBrief(cid);
-}
-export async function getBrief(cid: string): Promise<NextResponse> {
-  const fileEntry = await getFileEntry(cid);
-  if (fileEntry === null) {
-    return NextResponse.json(
-      { message: "Chart ID Not Found" },
-      { status: 404 }
-    );
-  }
-  return NextResponse.json({
-    ytId: fileEntry.ytId,
-    title: fileEntry.title,
-    composer: fileEntry.composer,
-    chartCreator: fileEntry.chartCreator,
-  } as ChartBrief);
 }
